--- src/controller/SearchTagsController.ts
@@ -1,17 +1,0 @@
-import {
-  Route,
-  Get,
-  Controller,
-  Security
-} from "tsoa";
-import { SearchTagsService } from "../service/SearchService";
-
-@Route("search")
-export class SearchTagsController extends Controller {
-  
-  @Get("tag") 
-  @Security("jwt", ["admin"])
-  public async getSearchTags() {
-    return await new SearchTagsService().getSearchTags();
-  }
-}
